Refactor session create response * Remove unnecessary interface * Use tenant name instead of tenant instance * Use session id instead of Session
