+++ configserver/src/main/java/com/yahoo/vespa/config/server/http/v2/SessionPrepareResponse.java
@@ -1,13 +1,12 @@
 // Copyright 2016 Yahoo Inc. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
 package com.yahoo.vespa.config.server.http.v2;
 
+import com.yahoo.config.provision.TenantName;
 import com.yahoo.container.jdisc.HttpRequest;
 import com.yahoo.slime.Slime;
 import com.yahoo.vespa.config.server.configchange.ConfigChangeActions;
 import com.yahoo.vespa.config.server.configchange.ConfigChangeActionsSlimeConverter;
 import com.yahoo.vespa.config.server.http.SessionResponse;
-import com.yahoo.vespa.config.server.session.Session;
-import com.yahoo.vespa.config.server.tenant.Tenant;
 
 /**
  * Creates a response for SessionPrepareHandler.
@@ -17,11 +16,12 @@ import com.yahoo.vespa.config.server.tenant.Tenant;
  */
 class SessionPrepareResponse extends SessionResponse {
 
-    public SessionPrepareResponse(Slime deployLog, Tenant tenant, HttpRequest request, Session session, ConfigChangeActions actions) {
+    public SessionPrepareResponse(Slime deployLog, TenantName tenantName, HttpRequest request, long sessionId, ConfigChangeActions actions) {
         super(deployLog, deployLog.get());
-        String message = "Session " + session.getSessionId() + " for tenant '" + tenant.getName() + "' prepared.";
-        this.root.setString("tenant", tenant.getName().value());
-        this.root.setString("activate", "http://" + request.getHost() + ":" + request.getPort() + "/application/v2/tenant/" + tenant.getName() + "/session/" + session.getSessionId() + "/active");
+        String message = "Session " + sessionId + " for tenant '" + tenantName.value() + "' prepared.";
+        this.root.setString("tenant", tenantName.value());
+        this.root.setString("activate", "http://" + request.getHost() + ":" + request.getPort() +
+                "/application/v2/tenant/" + tenantName.value() + "/session/" + sessionId + "/active");
         root.setString("message", message);
         new ConfigChangeActionsSlimeConverter(actions).toSlime(root);
     }
